chore: Move menu components to a subfolder
