+++ src/widgets/Menu/Menu.tsx
@@ -4,12 +4,12 @@ import throttle from "lodash/throttle";
 import Overlay from "../../components/Overlay/Overlay";
 import Flex from "../../components/Box/Flex";
 import { useMatchBreakpoints } from "../../hooks";
-import Logo from "./Logo";
-import Panel from "./Panel";
-import UserBlock from "./UserBlock";
+import Logo from "./components/Logo";
+import Panel from "./components/Panel";
+import UserBlock from "./components/UserBlock";
 import { NavProps } from "./types";
+import Avatar from "./components/Avatar";
 import { MENU_HEIGHT, SIDEBAR_WIDTH_REDUCED, SIDEBAR_WIDTH_FULL } from "./config";
-import Avatar from "./Avatar";
 
 const Wrapper = styled.div`
   position: relative;
