+++ src/widgets/Menu/components/Accordion.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import { MENU_ENTRY_HEIGHT } from "./config";
+import { MENU_ENTRY_HEIGHT } from "../config";
 import { MenuEntry, LinkLabel } from "./MenuEntry";
-import { PushedProps } from "./types";
-import { ArrowDropDownIcon, ArrowDropUpIcon } from "../../components/Svg";
+import { PushedProps } from "../types";
+import { ArrowDropDownIcon, ArrowDropUpIcon } from "../../../components/Svg";
 
 interface Props extends PushedProps {
   label: string;
