+++ src/widgets/Menu/components/Avatar.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
-import { Profile } from "./types";
-import NoProfileAvatar from "../../components/Svg/Icons/NoProfileAvatar";
+import { Profile } from "../types";
+import NoProfileAvatar from "../../../components/Svg/Icons/NoProfileAvatar";
 
 interface AvatarProps {
   profile: Profile;
