+++ src/widgets/Menu/components/Panel.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import styled from "styled-components";
 import PanelBody from "./PanelBody";
 import PanelFooter from "./PanelFooter";
-import { SIDEBAR_WIDTH_REDUCED, SIDEBAR_WIDTH_FULL } from "./config";
-import { PanelProps, PushedProps } from "./types";
+import { SIDEBAR_WIDTH_REDUCED, SIDEBAR_WIDTH_FULL } from "../config";
+import { PanelProps, PushedProps } from "../types";
 
 interface Props extends PanelProps, PushedProps {
   showMenu: boolean;
